refactor(wizard): deduplicate SSRF-guarded dispatch into shared trait
ConfigurationGenerator and ModelDiscovery carried an identical 69-line
block (HTTP-client test seam, setHttpClient(), dispatch(), error
sanitizing) that tripped the SonarCloud new-code duplication gate
(18.6% > 3%) and was flagged by review. Extract it into
SecureHttpDispatchTrait + the shared ErrorMessageSanitizerTrait; the
per-class vault audit reason stays a class constant.

The host-gate rejection now throws the dedicated
HostNotAllowedException instead of a bare RuntimeException
(SonarCloud php:S112), excluded from DI like the sibling exception
namespaces.

The test seam stays setter-based on purpose: TYPO3 core aliases
Psr\Http\Client\ClientInterface to GuzzleHttp\Client (public), so an
optional autowired constructor argument would inject the plain Guzzle
client in production and silently bypass the vault secure client and
its SSRF host gate.

Use https:// in the SSRF-guard test fixtures — the host gate fires on
the host, not the scheme — so the fixtures stop raising the
php:S5332 security hotspots that failed the hotspots-reviewed gate.
