+++ Classes/Service/SetupWizard/ModelDiscovery.php
@@ -11,15 +11,12 @@ namespace Netresearch\NrLlm\Service\SetupWizard;
 
 use Netresearch\NrLlm\Service\SetupWizard\DTO\DetectedProvider;
 use Netresearch\NrLlm\Service\SetupWizard\DTO\DiscoveredModel;
+use Netresearch\NrLlm\Utility\ErrorMessageSanitizerTrait;
 use Netresearch\NrVault\Http\SecureHttpClientFactory;
 use Netresearch\NrVault\Service\VaultServiceInterface;
-use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Log\LoggerInterface;
-use RuntimeException;
 use Throwable;
 
 /**
@@ -37,11 +34,14 @@ use Throwable;
  */
 final class ModelDiscovery implements ModelDiscoveryInterface
 {
+    use ErrorMessageSanitizerTrait;
+    use SecureHttpDispatchTrait;
+
     /**
-     * Test seam: when set, requests go through this client instead of the
-     * vault secure client. Production never sets it.
+     * Audit-log reason recorded by the vault secure client for every
+     * outbound request (see `SecureHttpDispatchTrait::dispatch()`).
      */
-    private ?ClientInterface $configuredHttpClient = null;
+    private const VAULT_DISPATCH_REASON = 'LLM setup-wizard model discovery';
 
     public function __construct(
         private readonly VaultServiceInterface $vault,
@@ -50,66 +50,6 @@ final class ModelDiscovery implements ModelDiscoveryInterface
         private readonly StreamFactoryInterface $streamFactory,
         private readonly LoggerInterface $logger,
     ) {}
-
-    /**
-     * Inject a custom HTTP client, bypassing the vault secure client.
-     *
-     * @internal Test seam only — production always dispatches through the
-     *           audited vault client (see `dispatch()`).
-     */
-    public function setHttpClient(ClientInterface $client): void
-    {
-        $this->configuredHttpClient = $client;
-    }
-
-    /**
-     * Send a request through the SSRF-guarded vault secure client.
-     *
-     * The host is gated up front via `isHostAllowed()` so a disallowed /
-     * private-range target is rejected before any secret-bearing header is
-     * sent; the vault client re-checks the host and validates the scheme as
-     * defence in depth.
-     *
-     * @throws Throwable when the host is disallowed or the request fails
-     */
-    private function dispatch(RequestInterface $request): ResponseInterface
-    {
-        // Test seam: an injected client bypasses the vault path entirely so
-        // unit tests can assert on the request the wizard built without hitting
-        // DNS or the host allowlist.
-        if ($this->configuredHttpClient !== null) {
-            return $this->configuredHttpClient->sendRequest($request);
-        }
-
-        // Reject a disallowed / private-range target up front, before any
-        // secret-bearing header reaches the wire. The vault client re-checks
-        // the host and validates the scheme inside sendRequest() as defence in
-        // depth, but failing here yields a clear, typed rejection.
-        $host = $request->getUri()->getHost();
-        if (!$this->httpClientFactory->isHostAllowed($host)) {
-            throw new RuntimeException(
-                sprintf('Host "%s" is not in the allowed hosts list', $host),
-                7438190001,
-            );
-        }
-
-        return $this->vault->http()
-            ->withReason('LLM setup-wizard model discovery')
-            ->sendRequest($request);
-    }
-
-    /**
-     * Strip secret-bearing query parameters from a message before it is
-     * returned to the client or logged. Mirrors `AbstractProvider`.
-     */
-    private function sanitizeErrorMessage(string $message): string
-    {
-        return (string)preg_replace(
-            '/([?&])(key|api_key|apikey|token|secret|access_token)=[^&\s]+/i',
-            '$1$2=***',
-            $message,
-        );
-    }
 
     /**
      * Test connection to provider.
