+++ Configuration/Services.yaml
@@ -20,6 +20,7 @@ services:
       - '../Classes/Attribute/*'
       - '../Classes/Domain/Model/*'
       - '../Classes/Provider/Exception/*'
+      - '../Classes/Service/SetupWizard/Exception/*'
       - '../Classes/Specialized/Exception/*'
       - '../Classes/Widgets/*'
